[infer][java] remove the hack around virtual calls on arrays
Summary:
This approach was requiring the `InferArray` class to always be part of the classpath, and the only benefit was to preserve the length of the arrays, when known, on calls to `clone()` method. However, adding it to the models would create circular dependencies between the models, the builtins and the tests.

The code is now simpler and we can more aggressively fail when classes that are supposed to be found from the classpath are not found.

Reviewed By: sblackshear

Differential Revision: D5703173

fbshipit-source-id: 3e6cea5