[infer][biabduction] no longer exchange the expressions associated to an attribute
Summary: It is not clear to me what the removed code was for in the first place. Basically, it was replacing the pure part of propositions in a semantically equivalent way, e.g. replacing `a = b /\ Attribute(a)` to `a = b /\ Attribute(b)`, and  `a = b /\ Attribute(b)` to `a = b /\ Attribute(a)`.

Reviewed By: jberdine

Differential Revision: D5657366

fbshipit-source-id: 93cd9e0