[infer][java] add the bootclasspath command line option to the genrule tests
Summary: Without this option, the frontend would fail to retrieve the type definition for the classes in the standard library, e.g. java.lang.String. Those failures are currently swallowed but the plan is to fail hard whenever the translation needs the definition of a class and is unable to find it as this should generally caused by some intergation bug with the build system.

Reviewed By: sblackshear

Differential Revision: D5770061

fbshipit-source-id: 20bad8f