[clang] fix evaluation order bug discovered by building with OCaml 4.05.0
Summary:
The behaviour of infer was observed to be different in 4.04.2 vs (4.05.0 or
4.04.2+flambda). Investigating further, the behaviour is also different in byte
vs native versions of infer under 4.04.2. Looking at the Changelog of 4.05.0
(thanks mbouaziz), there are fixes for inconsistent behaviours between byte
and native relative to evaluation order.

Lots of debugging later, this 1 line patch is born. Also use `List.concat_map`
instead of re-implementing it.

Reviewed By: jberdine

Differential Revision: D5793809

fbshipit-source-id: 374fb4c