[driver] remove infer-clang fake command
Summary:
"Running as clang" was its own infer subcommand. That's not terribly good
because it makes it hard to specify another subcommand, and in particular it
broke the `compile` subcommand in some integrations because "running as clang"
would always do capture. The cmake tests required to run with `--keep-going`
because of this.

Instead of having its own fake subcommand, simply add a new boolean in the
config for "infer runs as clang", as we do for javac already (used in the mvn
integration).

Also make logging of the environment better.

Reviewed By: jberdine

Differential Revision: D5813986

fbshipit-source-id: 72b96cd