[checkers] move aggregation by class out of cluster checker callbacks
Summary:
A Java cluster checker currently defines a "cluster" as all of the procedures in the same class.
But the cluster checker actually knows about all the procedures defined in the same source file.
In some checkers (such as thread-safety), we want to aggregate results across classes in the same file, not just methods in the same class.
This refactoring leaves the behavior the same for now, but will make it easier to do this in the near future.

Reviewed By: jeremydubreil

Differential Revision: D5885896

fbshipit-source-id: 0815fca