[log] Even better exception reraising
Summary:
`reraise` was error-prone when one forgot to save the backtrace between where the exception is caught and where it is reraised.
If any exception was raised (even caught) in between, the printed backtrace would be the one of the last exception thrown and it would be very confusing.

This diff kills `reraise` and introduces `reraise_after exn ~f` and `reraise_if exn ~f` to be used right after catching the exception.

Also turned some of them to the common pattern `try_finally ~f ~finally`.

Reviewed By: jvillard

Differential Revision: D5911244

fbshipit-source-id: 9883d1e