[siof] one access per sink, better report deduplication
Summary:
The previous domain for SIOF was duplicating some work with the generic Trace
domain, and basically was a bit confused and confusing. A sink was a set of
global accesses, and a state contains a set of sinks. Then the checker has to
needlessly jump through hoops to normalize this set of sets of accesses into a
set of accesses.

The new domain has one sink = one access, as suggested by sblackshear. This simplifies
a few things, and makes the dedup logic much easier: just grab the first report
of the list of reports for a function.

We only report on the fake procedures generated to initialise a global, and the
filtering means that we keep only one report per global.

Reviewed By: sblackshear

Differential Revision: D5932138

fbshipit-source-id: acb7285