[clang] do not translate dependent types
Summary:
Attempting to translate these will not go well as the declaration still depends
on some template arguments. Added a test that was previously crashing the
frontend.

Also extend the catching of "Unimplemented" and other errors to `translate_one_decl` as it was useful to debug this issue. In particular, reraise all exceptions and log some additional context when doing so.

update-submodule: facebook-clang-plugins

Reviewed By: mbouaziz

Differential Revision: D5976357

fbshipit-source-id: fca8e38