[inferbo] Do not report subsumed issues
Summary:
Next step to issue deduplication: do not keep safety conditions that are subsumed by others.
Only do it if they do not have infinite bound: replacing `0 < size` by `1 < size` is ok, but replacing it by `+oo < size` is not because it looks much more like a lack of precision.

Reviewed By: skcho

Differential Revision: D5978455

fbshipit-source-id: acc2384