[thread-safety] skip reporting on truncated traces
Summary:
Due to limitations in our Buck integration, the thread-safety analysis cannot create a trace that bottoms out in a Buck target that is not a direct dependency of the current target.
These truncated traces are confusing and tough to act on.
Until we can address these limitations, let's avoid reporting on truncated traces.

Reviewed By: jeremydubreil

Differential Revision: D5969840

fbshipit-source-id: 877b9de