[infer][biabduction] keep the files in the trace even when the exact location of the method is not known
Summary: The Java bytecode does not contain information about the location of abstract of interface methods. Before this diff, the analysis trace was tuncated and the file where the abstract or interface method was not included in the trace, which makes it harder to understand the Infer report, especially when the method is on a generated file that is not checked in the repository.

Reviewed By: sblackshear

Differential Revision: D6223612

fbshipit-source-id: c80c6f2