[siof] do not generate dummy derefs in initializer code
Summary:
This confuses the SIOF checker and causes false positives. This dummy deref is
generated for constructors of classes that are modeled as being pointer types
instead of the actual class in infer, typically for smart pointers. I do not
understand how this works.

The biabduction also analyses this code, so might now get confused itself.

Reviewed By: jberdine

Differential Revision: D6221817

fbshipit-source-id: 050c5a9