[SQLite] Set the synchronous flag to OFF
Summary:
On spinning disks the performance of commits are worse when `synchronous=NORMAL`.
Reading the documentation of SQLite, when `synchronous=OFF` there's a risk of DB corruption when the operating system crashes or the computer loses power before that data has been written to the disk surface; on the other hand, a crash in Infer should keep data in the DB in a sound state.
Buck reached the same conclusions too: https://github.com/facebook/buck/commit/468016227960ad73779334e995c34f285ad335ae

Reviewed By: mbouaziz

Differential Revision: D6413384

fbshipit-source-id: 99e4650