[infer] make sure adding or removing parameters does not affect the bug hash on Objective C methods
Summary: As Dulma pointed out, adding or removing paramters in a method in Objective C is changing the name of the method. Such changes should not make pre-exisiting issues reported as introduced. This diff is to prevent this by only keeping in the bug hash the part of the name that is before the first colon.

Reviewed By: dulmarod

Differential Revision: D6491215

fbshipit-source-id: 3c00fae