[racerd] Report only once per unique pair of final trace locations
Summary:
This diff adds a layer of report deduplication logic in addition to
the existing scheme.

Suppose issue 1 with trace1a and trace1b, and issue 2 with trace2a and
trace2b. If trace1a ends at the same location as trace2a (resp.,
trace2b) and trace1b ends at the same location as trace2b (resp.,
trace2a), then consider issues 1 and 2 to be duplicates.

This chooses to report the issue with the smaller sum of trace
lengths, breaking ties using the issue hashes, and eventually the
entire issue. Therefore there is a potential for flakiness with
respect the the choice of which report to make within a
hash-equivalence class.

Reviewed By: sblackshear

Differential Revision: D6519607

fbshipit-source-id: 63210ab