[exe_env] do not record the call graph in the execution environment
Summary:
Last piece of significant logic relying on the call graph: the execution
environment. Mostly carve out the logic to detect duplicate symbols to not rely
on the call graph.

Also make the keys of the `file_map` be source files, because not having the
"cg filename" makes it harder otherwise.

Reviewed By: sblackshear

Differential Revision: D6621196

fbshipit-source-id: cab50ba