[tenv] change interface to take `SourceFile.t` instead of `DB.filename`
Summary:
Also, make it explicit when we load the global tenv instead of the per-file tenv.

This allows for some nice simplifications in some places, notably:
- `tenv_file` is gone from `Exe_env.file_data`
- `DB.global_tenv_fname` is no more

This will help moving the tenv from the capture/source_file/ directories on the
filesystem to the database, as keys for the relevant table are `SourceFile.t`.

Reviewed By: mbouaziz

Differential Revision: D6796594

fbshipit-source-id: 1ffd5b0