[backend] Fix consistency of closures after bi-abduction and other executions.
Summary:
The captured variables of a closure are tuples (id, var, typ) with the implicit assumption
that &var -> id holds in the heap. This is true when the closure is created, but is not enforce otherwise.
This becomes a problem when the closure is stored in the heap, goes passed a bi-abduction, and then it's executed
(see new test). This was failing before this diff and now succeeds.

We add the verification of this constraint to the normalization of sigma.

At the moment I expect Precondition_not_met to be removed, but also later, we will be able to compute retain cycles
over the closures, as the correct captured variable info is kept through the execution.

Reviewed By: jvillard

Differential Revision: D6796525

fbshipit-source-id: a8a7655