[litho] move reporting of missing required props out of transfer functions
Summary:
This lets us fix the limitation of reporting false positives when a `private` function calls `build()` on a parameter without passing all of the required props.
We will now report such issues in the caller only if it fails to pass the required props.

An unfortunate consequence of this change is that we lose track of where the actual call to `build` occurs--we now report on the declaration of the caller function rather than on the call site of `build`.
I'll work on addressing that in a follow-up.

Reviewed By: jeremydubreil

Differential Revision: D6764153

fbshipit-source-id: 3b173e5