[clang] echo clang stdout when running non-plugin commands
Summary:
In some cases infer runs clang commands where we do not attach the plugin,
notably for pre-processor-only commands like `clang -E foo.c`. Usually these
commands are used by build systems to test their output, so infer should not
swallow it.

This makes infer echo the stdout of clang whenever we do not attach the plugin
to clang. When the plugin is attached, stdout is captured because that's where
the plugin outputs its results, so we cannot do the same in this case (not that
we would want to yet).

Fixes #696.

Reviewed By: mbouaziz

Differential Revision: D6912101

fbshipit-source-id: c4ad2e4